Informed of a better way to handle money in to be exact, which is decimal instead of float. The more you know.
All cash doubles changed to decimals.

+++ LemonadeStand/LemonadeStand/LemonadeStand.cs
@@ -7,7 +7,7 @@ namespace LemonadeStand
 {
     public class LemonadeStand
     {
-        internal double pricePerCup = 0;
+        internal decimal pricePerCup = 0.0m;
         LemonadeRecipe standRecipe = new LemonadeRecipe();
